+++ src/main/java/com/quasiris/qsf/query/FilterType.java
@@ -5,7 +5,7 @@ package com.quasiris.qsf.query;
  */
 public enum FilterType {
 
-    TERM("term"),MATCH("match"),MATCH_PHRASE("match_phrase"),RANGE("range"),SLIDER("slider");
+    TERM("term"),MATCH("match"),MATCH_PHRASE("match_phrase"),RANGE("range"),SLIDER("slider"),TREE("tree");
 
 
     private final String code;